requirements.txt
daemonize==2.4.2
pika==0.10.0
pycrypto==2.6.1
wsgiref==0.1.2
